[REFACT] reservation 디렉토리와 review 디렉토리 분리
