+++ src/main/java/com/cuk/catsnap/domain/photographer/entity/Photographer.java
@@ -5,7 +5,7 @@ import com.cuk.catsnap.domain.feed.entity.FeedComment;
 import com.cuk.catsnap.domain.feed.entity.FeedLike;
 import com.cuk.catsnap.domain.notification.entity.Notification;
 import com.cuk.catsnap.domain.reservation.entity.Reservation;
-import com.cuk.catsnap.domain.reservation.entity.Review;
+import com.cuk.catsnap.domain.review.entity.Review;
 import jakarta.persistence.*;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
