+++ src/main/java/com/cuk/catsnap/domain/reservation/entity/Reservation.java
@@ -3,6 +3,7 @@ package com.cuk.catsnap.domain.reservation.entity;
 import com.cuk.catsnap.domain.member.entity.Member;
 import com.cuk.catsnap.domain.notification.entity.ReservationNotification;
 import com.cuk.catsnap.domain.photographer.entity.Photographer;
+import com.cuk.catsnap.domain.review.entity.Review;
 import jakarta.persistence.*;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
