+++ src/main/java/com/cuk/catsnap/domain/review/entity/ReviewLike.java
@@ -1,5 +1,6 @@
-package com.cuk.catsnap.domain.reservation.entity;
+package com.cuk.catsnap.domain.review.entity;
 
+import com.cuk.catsnap.domain.member.entity.Member;
 import jakarta.persistence.*;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
@@ -7,21 +8,22 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 
 @Entity
-@Table(name="review_photo")
+@Table(name="review_like")
 @Getter
 @Builder
 @NoArgsConstructor
 @AllArgsConstructor
-public class ReviewPhoto {
-
+public class ReviewLike {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name="review_photo_id")
+    @Column(name="review_like_id")
     private Long id;
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name="review_id")
     private Review review;
 
-    private String PhotoUrl;
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name="member_id")
+    private Member member;
 }
