+++ src/main/java/com/cuk/catsnap/domain/review/entity/ReviewPhoto.java
@@ -1,6 +1,5 @@
-package com.cuk.catsnap.domain.reservation.entity;
+package com.cuk.catsnap.domain.review.entity;
 
-import com.cuk.catsnap.domain.member.entity.Member;
 import jakarta.persistence.*;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
@@ -8,22 +7,21 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 
 @Entity
-@Table(name="review_like")
+@Table(name="review_photo")
 @Getter
 @Builder
 @NoArgsConstructor
 @AllArgsConstructor
-public class ReviewLike {
+public class ReviewPhoto {
+
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name="review_like_id")
+    @Column(name="review_photo_id")
     private Long id;
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name="review_id")
     private Review review;
 
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name="member_id")
-    private Member member;
+    private String PhotoUrl;
 }
